refactor hook out of util and add it straight to Results component
