+++ app/src/organisms/PipetteWizardFlows/Results.tsx
@@ -9,6 +9,7 @@ import {
 } from '@opentrons/components'
 import { NINETY_SIX_CHANNEL } from '@opentrons/shared-data'
 import { SimpleWizardBody } from '../../molecules/SimpleWizardBody'
+import { useAttachedPipetteCalibrations } from '../Devices/hooks'
 import { SmallButton } from '../../atoms/buttons/OnDeviceDisplay'
 import { CheckPipetteButton } from './CheckPipetteButton'
 import { FLOWS } from './constants'
@@ -35,7 +36,6 @@ export const Results = (props: ResultsProps): JSX.Element => {
     isOnDevice,
     isFetching,
     setFetching,
-    recalibrate,
   } = props
   const { t, i18n } = useTranslation(['pipette_wizard_flows', 'shared'])
   const [numberOfTryAgains, setNumberOfTryAgains] = React.useState<number>(0)
@@ -43,6 +43,8 @@ export const Results = (props: ResultsProps): JSX.Element => {
     attachedPipettes[mount] != null
       ? attachedPipettes[mount]?.modelSpecs.displayName
       : ''
+  const pipCalibrationsByMount = useAttachedPipetteCalibrations()
+  const hasCalData = pipCalibrationsByMount[mount].offset?.lastModified != null
   let header: string = 'unknown results screen'
   let iconColor: string = COLORS.successEnabled
   let isSuccess: boolean = true
@@ -50,7 +52,7 @@ export const Results = (props: ResultsProps): JSX.Element => {
   let subHeader
   switch (flowType) {
     case FLOWS.CALIBRATE: {
-      header = t(recalibrate ? 'pip_recal_success' : 'pip_cal_success', {
+      header = t(hasCalData ? 'pip_recal_success' : 'pip_cal_success', {
         pipetteName: pipetteName,
       })
       break
