+++ app/src/organisms/PipetteWizardFlows/getPipetteWizardSteps.ts
@@ -1,6 +1,5 @@
 import { FLOWS, SECTIONS } from './constants'
 import { SINGLE_MOUNT_PIPETTES, LEFT, RIGHT } from '@opentrons/shared-data'
-import { useAttachedPipetteCalibrations } from '../Devices/hooks'
 import type {
   PipetteWizardStep,
   PipetteWizardFlow,
@@ -16,8 +15,6 @@ export const getPipetteWizardSteps = (
   isGantryEmpty: boolean,
   attachedPipettes: AttachedPipettesByMount
 ): PipetteWizardStep[] => {
-  const pipCalibrationsByMount = useAttachedPipetteCalibrations()
-  const hasCalData = pipCalibrationsByMount[mount].offset?.lastModified != null
   switch (flowType) {
     case FLOWS.CALIBRATE: {
       return [
@@ -32,7 +29,6 @@ export const getPipetteWizardSteps = (
           section: SECTIONS.RESULTS,
           mount: mount,
           flowType: flowType,
-          recalibrate: hasCalData,
         },
       ]
     }
