+++ app/src/organisms/PipetteWizardFlows/index.tsx
@@ -253,9 +253,6 @@ export const PipetteWizardFlows = (
         totalStepCount={totalStepCount}
         isFetching={isFetchingPipettes}
         setFetching={setIsFetchingPipettes}
-        recalibrate={
-          currentStep.recalibrate != null && currentStep.recalibrate === true
-        }
       />
     )
   } else if (currentStep.section === SECTIONS.MOUNT_PIPETTE) {
